Crate conrod_rendy

Structs

SimpleUiAux
A simple, provided implementation of the UiAux trait.

UiPipeline

UiPipelineDesc

UiTexture
A simple type that wraps a rendy Texture and provides a conrod_core::mesh::ImageDimensions implementation.

Constants

MODE_GEOMETRY
Ignore tex and draw simple, colored 2D geometry.

MODE_IMAGE
Draw an image from the texture at tex in the fragment shader.

MODE_TEXT
Draw text from the text cache texture tex in the fragment shader.

Traits

UiAux
Requirements for the auxiliary type within rendy graphs containing a UiPipeline node.