Struct conrod_rendy::SimpleUiAux

pub struct SimpleUiAux<B> where
    B: Backend,  {
    pub ui: Ui,
    pub image_map: ImageMap<UiTexture<B>>,
    pub dpi_factor: f64,
}

A simple, provided implementation of the UiAux trait.

Useful as a suitable auxiliary type when the UiPipeline is the only pipeline within the rendy graph.

Fields
ui: Uiimage_map: ImageMap<UiTexture<B>>dpi_factor: f64