Trait conrod_rendy::UiAux

pub trait UiAux {
    type Backend: Backend;
    fn ui(&self) -> &Ui;

    fn image_map(&self) -> &ImageMap<UiTexture<Self::Backend>>;

    fn dpi_factor(&self) -> f64;
}

Requirements for the auxiliary type within rendy graphs containing a UiPipeline node.

Associated Types

type Backend: Backend

Loading content...
Required methods

fn ui(&self) -> &Ui

The user interface to be drawn.

fn image_map(&self) -> &ImageMap<UiTexture<Self::Backend>>

Access to the user's images.

fn dpi_factor(&self) -> f64

The DPI factor for translating from conrod's pixel-agnostic coordinates to pixel coordinates for the underlying surface.

Loading content...